Refactored PatientModal - Moved functions and webcam from page into PatientModal component - Renamed some functions for standardisation
