+++ components/registration/PatientModal.js
@@ -1,7 +1,7 @@
-import Modal from 'react-modal';
-import CustomModal from '../CustomModal';
 import { Button, InputField, InputBox } from '@/components/TextComponents';
 import { venueOptions } from '@/utils/constants';
+import AppWebcam from '@/utils/webcam';
+import React, { useState } from 'react';
 
 const VenueOptions = ({ handleInputChange }) => (
   <div>
@@ -30,23 +30,33 @@ const VenueOptions = ({ handleInputChange }) => (
 );
 
 export function PatientModal({
-  modalIsOpen,
   formDetails,
   imageDetails,
-  cameraIsOpen,
-  renderWebcam,
   closeModal,
   handleInputChange,
   submitNewPatient,
-  toggleCameraOpen,
-  loading,
+  setImageDetails,
+  loading, //is this used?
 }) {
+  const [cameraIsOpen, setCameraIsOpen] = useState(false);
+  const [webcam, setWebcam] = useState(null);
+
+  const webcamSetRef = webcam => {
+    setWebcam(webcam);
+  };
+
+  const toggleCameraOpen = () => {
+    setCameraIsOpen(!cameraIsOpen);
+  };
+
+  const webcamCapture = () => {
+    const imageSrc = webcam.getScreenshot();
+    setImageDetails(imageSrc);
+    setCameraIsOpen(false);
+  };
+
   return (
-    <CustomModal
-      isOpen={modalIsOpen}
-      onRequestClose={closeModal}
-      showCloseButton={false}
-    >
+    <>
       <div className="grid grid-cols-2 gap-4">
         <InputField
           name="name"
@@ -114,7 +124,14 @@ export function PatientModal({
             </div>
           )}
 
-          {cameraIsOpen && <div>{renderWebcam()}</div>}
+          {cameraIsOpen && (
+            <div>
+              <AppWebcam
+                webcamSetRef={webcamSetRef}
+                webcamCapture={webcamCapture}
+              />
+            </div>
+          )}
           <div className="flex items-center justify-center mt-2">
             {cameraIsOpen ? (
               <Button colour="red" text="Cancel" onClick={toggleCameraOpen} />
@@ -133,6 +150,6 @@ export function PatientModal({
         <Button colour="green" text="Submit" onClick={submitNewPatient} />
         <Button colour="red" text="Close" onClick={closeModal} />
       </div>
-    </CustomModal>
+    </>
   );
 }
